refactor: add PilatesProfileSection component to modularize Pilates-related form fields in EditFormDialog
